Rename SegmenationMaskCollection to BinaryMaskCollection
We will use BinaryMaskCollection for other useful purposes, where it may not necessarily be the product of segmentation.

Test plan: travis.
Depends on #1610
