+++ starfish/core/binary_mask/__init__.py
@@ -1,0 +1,1 @@
+from .binary_mask import BinaryMaskCollection
