+++ starfish/core/binary_mask/expand.py
