+++ starfish/core/binary_mask/test/test_binary_mask.py
@@ -5,10 +5,10 @@ import pytest
 import xarray as xr
 
 from starfish.core.types import Axes, Coordinates
-from ..segmentation_mask import _validate_segmentation_mask, SegmentationMaskCollection
+from ..binary_mask import _validate_binary_mask, BinaryMaskCollection
 
 
-def test_validate_segmentation_mask():
+def test_validate_binary_mask():
     good = xr.DataArray([[True, False, False],
                          [False, True, True]],
                         dims=('y', 'x'),
@@ -16,7 +16,7 @@ def test_validate_segmentation_mask():
                                     y=[0, 1],
                                     xc=('x', [0.5, 1.5, 2.5]),
                                     yc=('y', [0.5, 1.5])))
-    _validate_segmentation_mask(good)
+    _validate_binary_mask(good)
 
     good = xr.DataArray([[[True], [False], [False]],
                          [[False], [True], [True]]],
@@ -27,7 +27,7 @@ def test_validate_segmentation_mask():
                                     zc=('z', [0.5, 1.5]),
                                     yc=('y', [1.5, 2.5, 3.5]),
                                     xc=('x', [42.5])))
-    _validate_segmentation_mask(good)
+    _validate_binary_mask(good)
 
     bad = xr.DataArray([[1, 2, 3],
                         [4, 5, 6]],
@@ -37,14 +37,14 @@ def test_validate_segmentation_mask():
                                    xc=('x', [0.5, 1.5, 2.5]),
                                    yc=('y', [0.5, 1.5])))
     with pytest.raises(TypeError):
-        _validate_segmentation_mask(bad)
+        _validate_binary_mask(bad)
 
     bad = xr.DataArray([True],
                        dims=('x'),
                        coords=dict(x=[0],
                                    xc=('x', [0.5])))
     with pytest.raises(TypeError):
-        _validate_segmentation_mask(bad)
+        _validate_binary_mask(bad)
 
     bad = xr.DataArray([[True]],
                        dims=('z', 'y'),
@@ -53,12 +53,12 @@ def test_validate_segmentation_mask():
                                    zc=('z', [0.5]),
                                    yc=('y', [0.5])))
     with pytest.raises(TypeError):
-        _validate_segmentation_mask(bad)
+        _validate_binary_mask(bad)
 
     bad = xr.DataArray([[True]],
                        dims=('x', 'y'))
     with pytest.raises(TypeError):
-        _validate_segmentation_mask(bad)
+        _validate_binary_mask(bad)
 
 
 def test_from_label_image():
@@ -70,7 +70,7 @@ def test_from_label_image():
     physical_ticks = {Coordinates.Y: [1.2, 2.4, 3.6, 4.8, 6.0],
                       Coordinates.X: [7.2, 8.4, 9.6, 10.8, 12]}
 
-    masks = list(SegmentationMaskCollection.from_label_image(
+    masks = list(BinaryMaskCollection.from_label_image(
         label_image, physical_ticks).masks())
 
     assert len(masks) == 2
@@ -112,8 +112,8 @@ def test_to_label_image():
     physical_ticks = {Coordinates.Y: [1.2, 2.4, 3.6, 4.8, 6.0],
                       Coordinates.X: [7.2, 8.4, 9.6, 10.8, 12, 15.5]}
 
-    masks = SegmentationMaskCollection.from_label_image(label_image,
-                                                        physical_ticks)
+    masks = BinaryMaskCollection.from_label_image(label_image,
+                                                  physical_ticks)
 
     assert np.array_equal(masks.to_label_image(), label_image)
 
@@ -127,13 +127,13 @@ def test_save_load():
     physical_ticks = {Coordinates.Y: [1.2, 2.4, 3.6, 4.8, 6.0],
                       Coordinates.X: [7.2, 8.4, 9.6, 10.8, 12]}
 
-    masks = SegmentationMaskCollection.from_label_image(label_image,
-                                                        physical_ticks)
+    masks = BinaryMaskCollection.from_label_image(label_image,
+                                                  physical_ticks)
 
     path = 'data.tgz'
     try:
         masks.save(path)
-        masks2 = SegmentationMaskCollection.from_disk(path)
+        masks2 = BinaryMaskCollection.from_disk(path)
         for m, m2 in zip(masks.masks(), masks2.masks()):
             assert np.array_equal(m, m2)
     finally:
