+++ starfish/core/binary_mask/test/test_expand.py
