+++ starfish/core/binary_mask/util.py
